requirements.txt
apex==0.9.10dev
cuml_cu11==23.12.0
data_process==0.3.9
dlib==19.24.1
fvcore==0.1.5.post20221221
loralib==0.1.2
mmcv==2.1.0
mmdet==3.3.0
mmseg==1.3.0
model==0.6.0
numpy==1.23.5
omegaconf==2.3.0
opencv_contrib_python_headless==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==10.2.0
scikit_learn==1.2.2
scipy==1.7.1
setuptools==66.0.0
submitit==1.5.1
torch==2.0.0+cu117
torchmetrics==0.10.3
torchsummary==1.5.1
torchvision==0.15.0+cu117
xformers==0.0.23.post1
